Prepared first version of DB migration tool
Database migration tool carries out migration of databases from mysql to postgresql. You can easily check out the list of tables and columns. It also has a module to download mysql databases to XML file in a logical structure. Modules for migration of postgresql databases to mysql and upload of XML-based databases to mysql will be released asap.